devtools/server/tests/mochitest/test_memory_allocations_07.html

<!DOCTYPE HTML>
<html>
<!--
Bug 1192335 - Test getting the byte sizes for allocations.
-->
<head>
  <meta charset="utf-8">
  <title>Memory monitoring actor test</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
<pre id="test">
<script src="memory-helpers.js" type="application/javascript"></script>
<script>
"use strict";

window.onload = function () {
  SimpleTest.waitForExplicitFinish();

  Task.spawn(function* () {
    let { memory, client } = yield startServerAndGetSelectedTabMemory();
    yield memory.attach();

    let allocs = [];
    function allocator() {
      allocs.push(new Object());
    }

    yield memory.startRecordingAllocations();

    allocator();
    allocator();
    allocator();

    let response = yield memory.getAllocations();
    yield memory.stopRecordingAllocations();

    ok(response.allocationSizes, "The response should have bytesizes.");
    is(response.allocationSizes.length, response.allocations.length,
       "There should be a bytesize for every allocation.");
    ok(response.allocationSizes.length >= 3,
       "There are atleast 3 allocations.");
    ok(response.allocationSizes.every(isPositiveNumber),
       "every bytesize is a positive number");

    yield memory.detach();
    destroyServerAndFinish(client);
  });
};

function isPositiveNumber(n) {
  return typeof n === "number" && n > 0;
}
</script>
</pre>
</body>
</html>
